Improve logging

+++ test-include/FreeRTOSConfig.h
@@ -42,8 +42,8 @@
 #define configUSE_IDLE_HOOK 1
 #define configUSE_TICK_HOOK 1
 #define configUSE_DAEMON_TASK_STARTUP_HOOK 0
-#define configTICK_RATE_HZ (1000)                                     /* In this non-real time simulated environment the tick frequency has to be at least a multiple of the Win32 tick frequency, and therefore very slow. */
-#define configMINIMAL_STACK_SIZE ((unsigned short) PTHREAD_STACK_MIN) /* The stack size being passed is equal to the minimum stack size needed by pthread_create(). */
+#define configTICK_RATE_HZ (1000)
+#define configMINIMAL_STACK_SIZE (PTHREAD_STACK_MIN)
 #define configTOTAL_HEAP_SIZE ((size_t) (512 * 1024))
 #define configMAX_TASK_NAME_LEN (12)
 #define configUSE_TRACE_FACILITY 1
